streamlit_app/requirements.txt
streamlit==1.41.0
requests==2.32.3
Pillow==11.1.0
pandas==2.2.3
mistralai>=1.0.0
bcrypt==4.2.1
PyJWT==2.9.0
psycopg2-binary==2.9.10
plotly==5.24.1
openpyxl==3.1.5
reportlab==4.1.0
